+++ src/pages/AnswerPage.jsx
@@ -1,14 +1,13 @@
 import { useCallback, useEffect, useState } from "react";
-import { useNavigate, useParams } from "react-router";
-import { deleteSubject } from "../api/subjects";
+import { useParams } from "react-router";
 import { getQuestionList } from "../api/questions";
 import QuestionBox from "../components/QuestionBox";
-import Button from "../components/Button";
 import QuestionContainer from "../components/QuestionContainer";
 import QuestionList from "./../components/QuestionList";
 import useInfiniteScroll from "../hooks/useInfiniteScroll";
 import useSubject from "../hooks/useSubject";
 import useInitialQuestion from "../hooks/useInitialQuestion";
+import DeleteButton from "../components/DeleteButton";
 
 const AnswerPage = () => {
   const [questionList, setQuestionList] = useState([]);
@@ -16,8 +15,6 @@ const AnswerPage = () => {
   const [offset, setOffset] = useState(0);
   const [isMoreQuestion, setIsMoreQuestion] = useState(false);
   const [isLoading, setIsLoading] = useState(false);
-  const [isDeleting, setIsDeleting] = useState(false);
-  const navigate = useNavigate();
   const { subject, questionCount } = useSubject(id);
 
   const { isInitialLoading } = useInitialQuestion({
@@ -42,37 +39,12 @@ const AnswerPage = () => {
 
   const { ref } = useInfiniteScroll({ callback: getMoreData, isMoreQuestion });
 
-  const handleDelete = async () => {
-    if (isDeleting) return;
-    setIsDeleting(true);
-    try {
-      await deleteSubject({ subjectId: id });
-      const subjects = JSON.parse(localStorage.getItem("subjects") || "[]");
-      const updatedSubjects = subjects.filter(
-        (subject) => String(subject.id) !== String(id),
-      );
-      localStorage.setItem("subjects", JSON.stringify(updatedSubjects));
-      navigate("/list");
-    } catch {
-      alert("삭제를 실패했습니다.");
-    } finally {
-      setIsDeleting(false);
-    }
-  };
-
   useEffect(() => window.scrollTo(0, 0), []);
 
   return (
     <QuestionContainer subject={subject}>
       <div className="relative w-full max-w-716">
-        <Button
-          disabled={isDeleting}
-          type="round"
-          onClick={handleDelete}
-          className="pc:top-10 tablet:top-0 tablet:w-100 tablet:h-35 tablet:text-[15px] absolute top-23 right-0 w-70 text-[10px]/25 font-extralight"
-        >
-          {isDeleting ? "삭제중" : "삭제하기"}
-        </Button>
+        <DeleteButton id={id} />
         <QuestionBox count={questionCount}>
           <QuestionList
             isInitialLoading={isInitialLoading}
